refactored session1 page
